pyproject.toml
[tool.poetry]
name = "upower-stats"
version = "0.1.0"
description = ""
authors = ["Andrew Haigh <hello@nelf.in>"]

[tool.poetry.dependencies]
python = "^3.7"
dbus-python = "^1.2.16"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
